In the middle of adding pit scout textboxes
Not finished yet, in the middle of hard coding the auto path textbox

But there's an issue with both the ui and exporting to csv.

Ui is just overlapping, make sure the auto box textbox (named mMotorInfo3) is the textbox the rest of the spinners are coming beneath.

But, none of the motorInfos are returning any text to the csv files, even our og motorInfo which should have been working.